compiler: use int128 to represent integers
We used to represent integers using usize, and sometimes isize. That
introduced a lot of complexity and type guessing to avoid overflows.
This commit switches to i128 to represent all integer types. i128 should
be big enough to hold any value used by the RISC-V model, including
u64::MAX and i64::MIN.

+++ rv64/src/lib.rs
@@ -54,12 +54,12 @@ impl Core {
 
     /// Set the `pmpaddr<index>` register to the given value.
     pub fn set_pmpaddr(&mut self, index: usize, val: u64) {
-        raw::pmpWriteAddrReg(self, index, BitVector::new(val));
+        raw::pmpWriteAddrReg(self, index as i128, BitVector::new(val));
     }
 
     /// Set the `pmpcfg<index>` register to the given value.
     pub fn set_pmpcfg(&mut self, index: usize, val: u64) {
-        raw::pmpWriteCfgReg(self, index, BitVector::new(val));
+        raw::pmpWriteCfgReg(self, index as i128, BitVector::new(val));
     }
 
     /// Check if an 8 byte access is allowed with the current mode and PMP configuration.
@@ -269,7 +269,7 @@ pub const fn new_core(config: raw::Config) -> Core {
         htif_exit_code: BitVector::new(0),
         htif_cmd_write: false,
         htif_payload_writes: BitVector::new(0),
-        tlb: [DEFAULT_TLB_ENTRY; raw::num_tlb_entries],
+        tlb: [DEFAULT_TLB_ENTRY; raw::num_tlb_entries as usize],
         satp: BitVector::new(0),
         hart_state: raw::HartState::HART_ACTIVE(()),
         config,
